requirements.txt
colorlabels
python-twitter
twitterscraper
nltk
stop-words
pyldavis
gensim
python-decouple
